Move getTablesUsedInFrom() from ActiveQuery to Query
enable this feature on the Query level
